Owners can remove availability through the app
In order to implement this feature we realised we needed to add a route
for owners to update the details of their spaces. This in turn required
a route for owners to see only the spaces they own rather than a list of
all spaces owned by everyone, and additional work in the model.