Made consistency changes to class definitions
Rubocop was complaining about the length of some methods - and this was
in part due to how we'd spread our SQL and object.new references across
multiple lines. This commit shouldn't be merged if people think it's
just less readable and we should ignore rubocop - but the option is now
there just in case!